refactor: split startup state into a state for awaiting renewal and another for beginning the creation of a certificate
